User can only get Clients they have created

+++ controllers/clients.js
@@ -5,15 +5,31 @@ const Client = require('../models/client')
 const User = require('../models/user')
 
 clientsRouter.get('/', async (request, response) => {
+  const token = utils.getTokenFrom(request)
+  const decodedToken = jwt.verify(token, process.env.TOKEN_KEY)
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({
+      error: 'token missing or invalid'
+    })
+  }
+  console.log(decodedToken.id)
   const clients = await Client
-    .find({})
+    .find({ user: decodedToken.id })
     .populate('user', { username: 1, name: 1 })
   response.json(clients.map(client => client.toJSON()))
 })
 
 clientsRouter.get('/:id', async (request, response) => {
+  const token = utils.getTokenFrom(request)
+  const decodedToken = jwt.verify(token, process.env.TOKEN_KEY)
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({
+      error: 'token missing or invalid'
+    })
+  }
+  
   const client = await Client.findById(request.params.id)
-  if (client) {
+  if (client.user === decodedToken.id) {
     response.json(client.toJSON())
   } else {
     response.status(404).end()
@@ -47,6 +63,13 @@ clientsRouter.post('/', async (request, response) => {
 })
 
 clientsRouter.delete('/:id', async (request, response) => {
+  const token = utils.getTokenFrom(request)
+  const decodedToken = jwt.verify(token, process.env.TOKEN_KEY)
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({
+      error: 'token missing or invalid'
+    })
+  }
     await Client.findByIdAndRemove(request.params.id)
     response.status(204).end()
 })
